+++ src/datastream.h
@@ -4,13 +4,13 @@
 
 /* TODO: should point to a struct qr here */
 struct datastream_raw {
-	uint8_t		raw[QR_PAYLOAD_MAX];
-	int		bits;
+	uint8_t raw[QR_PAYLOAD_MAX];
+	size_t bits;
 };
 
 struct datastream_data {
-	uint8_t         data[QR_PAYLOAD_MAX];
-	int		bits;
+	uint8_t data[QR_PAYLOAD_MAX];
+	size_t bits;
 };
 
 void
@@ -21,7 +21,7 @@ read_data(const struct qr *q,
 	struct datastream_raw *ds);
 
 int
-take_bits(struct datastream_data *ds, int len, int *ds_ptr);
+take_bits(struct datastream_data *ds, size_t len, size_t *ds_ptr);
 
 #endif
